add twitter dependencies and check example
