+++ package.json
@@ -7,9 +7,8 @@
     "mongoose": "~4.2.0",
     "mongoose-auto-increment": "~2.1.3",
     "passport": "~0.1.17",
-    "passport-twitter": "~1.0.2",
     "ejs": "*",
-    "twitter-js-client": "*"
+    "twitter-node-client": "*"
   },
   "devDependencies": {
     "grunt": "*",
